Refactoring EC2 to return all instances rather than limiit of 10
