--- ava/config.py
@@ -1,12 +1,0 @@
-# Weather API
-WX_API_KEY = "<YOUR-API-KEY>"
-WX_LOCATION = "<YOUR-DEFAULT-LOCATION>"
-WX_METRIC_TEMP = "<celsius|fahrenheit|kelvin>"
-WX_METRIC_WIND = "<miles_hour|km_hour|knots>"
-
-# Github API
-GITHUB_USER = "<USERNAME>"
-GITHUB_PASS = "<PASSWORD>"
-
-# Tidy
-TIDY_ROOT = None  # "</path/to/root>"
